move getAnonTopicNullifier to passportInterface
