+++ apps/anon-message-client/next.config.js
@@ -1,6 +1,7 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   output: "export",
+  transpilePackages: ["@pcd/passport-interface"],
   webpack(config, { nextRuntime }) {
     // as of Next.js latest versions, the nextRuntime is preferred over `isServer`, because of edge-runtime
     if (typeof nextRuntime === "undefined") {
